split fully quantized to more files
