+++ model_compression_toolkit/core/common/user_info.py
@@ -31,6 +31,7 @@ class UserInformation:
         self.mixed_precision_cfg = None
         self.fusions = []
         self.final_kpi = None
+        self.custom_objects = {}
 
     def set_input_scale(self, scale_value: float):
         """
