+++ model_compression_toolkit/exporter/fully_quantized/keras/builder/fully_quantized_model_builder.py
@@ -1,4 +1,4 @@
-# Copyright 2022 Sony Semiconductor Israel, Inc. All rights reserved.
+# Copyright 2022 Sony Semiconductors Israel, Inc. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,41 +12,42 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
+
+from model_compression_toolkit.core.keras.quantizer.input_layer_quantize_transform import InputLayerWrapperTransform
+
 from typing import List, Tuple
 
 import tensorflow as tf
-import tensorflow_model_optimization.quantization.keras.graph_transformations.model_transformer as mt
 from keras.models import Model
+from tensorflow.python.util.object_identity import Reference as TFReference
+from tensorflow_model_optimization.python.core.quantization.keras.default_8bit.default_8bit_quantize_configs import \
+    NoOpQuantizeConfig
+from tensorflow_model_optimization.python.core.quantization.keras.quantize_wrapper import QuantizeWrapper
 
+from model_compression_toolkit.core import common
 from model_compression_toolkit.core.common import BaseNode
+from model_compression_toolkit.core.common.framework_info import FrameworkInfo
 from model_compression_toolkit.core.common.user_info import UserInformation
 from model_compression_toolkit.core.keras.back2framework.keras_model_builder import KerasModelBuilder, \
     is_layer_fake_quant, get_node_name_from_layer
-from model_compression_toolkit.core.keras.quantizer.input_layer_quantize_transform import \
-    InputLayerWrapperTransform
-
-# As from Tensorflow 2.6, keras is a separate package and some classes should be imported differently.
-from model_compression_toolkit.core.keras.quantizer.mixed_precision.selective_quantize_config import \
-    SelectiveQuantizeConfig
-
-if tf.__version__ < "2.6":
-    from tensorflow.python.keras.layers.core import TFOpLambda, SlicingOpLambda
-else:
-    from keras.layers.core import TFOpLambda, SlicingOpLambda
-
-from tensorflow_model_optimization.python.core.quantization.keras.quantize_wrapper import QuantizeWrapper
-from model_compression_toolkit.core.common.logger import Logger
-from model_compression_toolkit.core import common
-from model_compression_toolkit.core.common.framework_info import FrameworkInfo
 from model_compression_toolkit.core.keras.default_framework_info import DEFAULT_KERAS_INFO
-from model_compression_toolkit.core.keras.quantizer.mixed_precision.quantization_config_factory import \
-    quantization_config_builder_mixed_precision
-from tensorflow.python.util.object_identity import Reference as TFReference
+from model_compression_toolkit.exporter.fully_quantized.keras.builder.quantize_config_to_node import \
+    get_quantization_config
+from model_compression_toolkit.exporter.fully_quantized.keras.quantize_configs.activation_quantize_config import \
+    ActivationQuantizeConfig
+from model_compression_toolkit.exporter.fully_quantized.keras.quantize_configs.weights_activation_quantize_config import \
+    WeightsActivationQuantizeConfig
+from model_compression_toolkit.exporter.fully_quantized.keras.quantize_configs.weights_quantize_config import WeightsQuantizeConfig
+
+from model_compression_toolkit.exporter.fully_quantized.keras.quantizers.fq_quantizer import FakeQuantQuantizer
+from model_compression_toolkit.exporter.fully_quantized.keras.quantizers.uniform_quantizer import UniformQuantizer, \
+    WeightsUniformQuantizer
+import tensorflow_model_optimization.quantization.keras.graph_transformations.model_transformer as mt
 
 
-class MixedPrecisionKerasModelBuilder(KerasModelBuilder):
+class FullyQuantizedKerasModelBuilder(KerasModelBuilder):
     """
-    Builder of mixed-precision Keras models.
+    Builder of fully-quantized Keras models.
     """
 
     def __init__(self,
@@ -82,9 +83,6 @@ class MixedPrecisionKerasModelBuilder(KerasModelBuilder):
             Output of the node.
 
         """
-        if node.is_all_activation_candidates_equal():
-            # otherwise, we want to use the float tensor when building the model for MP search
-            return node.candidates_quantization_cfg[0].activation_quantization_cfg.quantize_node_output(input_tensors)
         return input_tensors
 
     def build_model(self) -> Tuple[Model, UserInformation]:
@@ -95,36 +93,29 @@ class MixedPrecisionKerasModelBuilder(KerasModelBuilder):
         """
         model, user_info = super().build_model()
 
-        conf_nodes_names = self.graph.get_configurable_sorted_nodes_names()
+        def _wrap_layer_with_quantize_config(layer):
 
-        def _quantize_multiple_nbits(layer):
             nodes = self.graph.find_node_by_name(get_node_name_from_layer(layer))
+
             if len(nodes) == 1:
                 node = nodes[0]
-                # Wrap only if its weights should be quantized
-                if node.name in conf_nodes_names:
-                    if node.layer_class in [TFOpLambda, SlicingOpLambda]:
-                        Logger.critical(f"Activation mixed-precision is not supported for layers of type "
-                                        f"{node.layer_class}. Please modify the TargetPlatformModel object, "
-                                        f"such that layers of type {node.layer_class} "
-                                        f"won't have more than one quantization configuration option.")
-                    return QuantizeWrapper(layer, quantization_config_builder_mixed_precision(node, self.fw_info))
-                return layer
+                return QuantizeWrapper(layer, get_quantization_config(node,
+                                                                      self.fw_info))
 
             elif is_layer_fake_quant(layer):
                 return layer
+
             else:
                 raise Exception(
                     f'Mismatch between keras model and graph cant find node named: '
                     f'{get_node_name_from_layer(layer)}')
 
-        # clone each layer in the model and apply _quantize to the layer.
+        # clone each layer in the model and apply _wrap_layer_with_quantize_config to the layer.
         model = tf.keras.models.clone_model(model,
                                             input_tensors=None,
-                                            clone_function=_quantize_multiple_nbits)
+                                            clone_function=_wrap_layer_with_quantize_config)
 
-        # We use a model transformer to wrap the input layer with QuantizeWrapper,
-        # to allow layer configuration to different bitwidths.
+        # We use a model transformer to wrap the input layer with QuantizeWrapper.
         # A model transformer allows to modify a layer in an existing model, by applying the given list of
         # transformers on the model (in this case,
         # we only apply single transformer - InputLayerQuantizeTransform)
@@ -132,15 +123,28 @@ class MixedPrecisionKerasModelBuilder(KerasModelBuilder):
 
         input_transformer = mt.ModelTransformer(model, [InputLayerWrapperTransform(inp,
                                                                                    self.fw_info,
-                                                                                   quantization_config_builder_mixed_precision(inp, self.fw_info),
-                                                                                   self.get_custom_objects())
+                                                                                   get_quantization_config(inp,
+                                                                                                           self.fw_info),
+                                                                                   self.get_custom_objects()
+                                                                                   )
                                                         for inp in model_inputs])
         model = input_transformer.transform()[0]
 
-        return model, user_info
+        user_info.custom_objects = self.get_custom_objects()
 
+        return model, user_info
 
     def get_custom_objects(self):
         return {"QuantizeWrapper":QuantizeWrapper,
-                "SelectiveQuantizeConfig": SelectiveQuantizeConfig}
+                "WeightsActivationQuantizeConfig": WeightsActivationQuantizeConfig,
+                "ActivationQuantizeConfig": ActivationQuantizeConfig,
+                "WeightsQuantizeConfig": WeightsQuantizeConfig,
+                "WeightsUniformQuantizer": WeightsUniformQuantizer,
+                "UniformQuantizer": UniformQuantizer,
+                "NoOpQuantizeConfig": NoOpQuantizeConfig,
+                "FakeQuantQuantizer": FakeQuantQuantizer}
+
+
+
+
